+++ CHANGELOG.md
@@ -49,6 +49,8 @@ Items that are related, such as breaking changes, new features, or changes to ex
 
 - [@siteimprove/alfa-selector](packages/alfa-selector): `Selector#equals()` avoids narrowing to `never` when used on a different `Selector`. ([#732](../../pull/732))
 
+- [@siteimprove/alfa-scraper](packages/alfa-scraper): `Scraper.scrape()` now correctly sets the viewport height to the specified device height instead of its width.
+
 ## [0.12.0](../../compare/v0.11.0...v0.12.0) (2021-02-26)
 
 ### Breaking